requirements.txt
torch>=2.1.0
tokenizers>=0.15.0
datasets>=2.16.0
wandb>=0.16.0
safetensors>=0.4.0
lm-eval>=0.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
